refactor: split off the displaying of images into helper function to keep display single listing simple
